[iOS] Animated transition from tab grid (selected tab not visible)
This is the animation when presenting a tab from the tab grid and the
current selected tab isn't visible (so the full grid zoom can't easily
be done). It would also be used when tapping 'done' while on the remote
tabs page.

This CL sets this to be the transition when presenting a tab in all
cases. A future CL will use the more detailed transition when it's
available.

The animation code is straightforward: set a zero alpha and scale
transform, then animate back to full alpha and the identity transform.

Bug: 804540
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I58f6ac505352f39ffc9f92e2dc3b85898944a2df
Reviewed-on: https://chromium-review.googlesource.com/956147
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#542447}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf6a1bec8aeafeba246857592d9743212d6164bc