[Nav Experiment] Fix "Untitled" tab after cold start.
WKBasedNavigationManager::Restore() calls WebState::LoadCurrentURL() to
load the restore_session.html into WKWebView. However, after a cold
start, LoadCurrentURL() early exits because there is no _containerView,
and the pending navigaiton to restore_session.html is only completed
when the tab becomes active. While a tab is in this state, the WebState
title is empty, causing "Untitled" to be displayed in the tab switcher,
and any call to BuildSessionStorage() will incorrectly save an empty
session. If the app is backgrounded, the previously restored session
will be lost.

This CL fixes the issue by temporarily caching the last restored session
storage in WebStateImpl, so it can be returned in GetTitle() and
BuildSessionStorage() in the intermediate state. The cached version is
invalidated as soon as a new navigation item is committed.

Bug: 822771
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I573c33f8d3282d4af0ef98c19fedb767c7ab987d
Reviewed-on: https://chromium-review.googlesource.com/969638
Commit-Queue: Danyao Wang <danyao@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#544205}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec3eea934bba1d47f717514fd75f1c27801f3c28