[iOS] Adding identity picker view
This patch adds a simple view that will be used to redo the sign-in workflow
for unity project.
This patch implements the IdentityPickerView which is the grey button with the
down arrow. This can be seen in the mock[1] for the sign-in workflow.

The accessibility is not done yet. The rest of the dialog will be implemented
soon.

IdentityPickerView:
https://drive.google.com/open?id=1bHM8QGHl7uTVMcUMLm-v3WRclj4kNyEc
IdentityPickerView in situe:
https://drive.google.com/open?id=1P3ZkJqNnk9ShIUwdaSbX-aZjvPn7NReg
IdentityPickerView with email only:
https://drive.google.com/open?id=18jD8IZlngQbKa_KkZ6ZADsxDdRl32yn4
IdentityPickerView with big fonts:
https://drive.google.com/open?id=1OLiDdQC6Tl1qBiG7ciVliAqX0LAwTXxz
IdentityPickerView email only with big font:
https://drive.google.com/open?id=1JkVtbTWQyCLwxwUeu0x_0nTKPvhFD9rE


Mock[1]:
https://docs.google.com/presentation/d/1cZfr5FGWGSy0PNaQ8uzik0alLAH-5glh1vsb030vha8/edit?ts=5aba5455#slide=id.g353bc22139_3_19

Bug: 827072
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Ia3caddb069b951e734161901325206555d8af0ae
Reviewed-on: https://chromium-review.googlesource.com/995417
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#548812}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 07c70e29cfeb2eaff53ded5a199c72ff4a47b46b