[iOS] Make GridConsumer idempotent
This CL makes the GridConsumer protocol not depend on the state
of the items list that it reports changes for. To do this, it only
uses list indexes as parameters for destinations (of insertions or
moves). In all other cases, the mutations specified by the protocol
are expressed in terms of the IDs of the items that are changed,
including the selected item.

GridViewController now keeps track of the ID of the selected item,
and determines (finds) its index dynamically.

Bug: 986379
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Id2465c9740e505be659126068bc2c0e0cbf8a8d3
Reviewed-on: https://chromium-review.googlesource.com/999712
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#549161}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ca5d408fe5cd422bc9dff8f6f9f707c597cb2963