Add Spotlighted state to the toolbar button
This CL adds a state to the toolbar button. This new state,
"Spotlighted" is used to put emphasis on a button in the toolbar.
It can be used to show that the page is bookmarked, or to indicate to
which button the popup menu displayed is linked (done in a future CL).

Bug: 817752
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I954cb0b26a577212ca8e3d8840b99f5be4aa1a2d
Reviewed-on: https://chromium-review.googlesource.com/1000785
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#550597}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 06ca303bc30849d0b3886d39e8fe58322f2c381f