[iOS] Adds scroll functions for UnifiedConsentViewController
The user needs to scroll to the end of the consent before to be able to
sign-in. This patch adds a test to know if the end of the user consent
is reached and a function to scroll to the end.

This is related to the "MORE" button:
https://drive.google.com/open?id=1B4zk-U0-rONMX9Ht_wC91jkMTMqDwVk6
https://drive.google.com/open?id=1d-TwNayGlC5ZxnwtemSTS4Jvrtwu7xLP

Related to:
crrev.com/c/1009906
crrev.com/c/1010346

Bug: 827072
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I0debff1607572ec27e22ffeb587aa652fc9db241
Reviewed-on: https://chromium-review.googlesource.com/1024833
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#554070}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d8a63647383c3369cc725ca0ede46a0b7c12d7a0