[ios] Adds MDCAppBar support to ChromeTableViewController.
Initially appbar support was added to SettingsRootTableViewController,
as that was the only view controller which needed it.  But now, we
expect to support MDCAppBar in bookmarks screens as well, so this CL
moves it to a common base class.  We follow the pattern set by
CollectionViewController and include an appBarStyle parameter which
determines whether or not an MDCAppBar is created.

BUG=839439,839450,839460

Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I057b15113987d91451ca3a104aba95dd17667cda
Reviewed-on: https://chromium-review.googlesource.com/1043006
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#555892}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d4ada0604de844a1c5206e5f17e3584ce83eec8