[ios] Replace recently added DCHECK with TODO to fix underlying issue
A DCHECK was recently added that revealed that the
tab grid cell is nil sometimes in |-cellForItemAtIndexPath:|.
The cell should never be nil in this method.
This is an issue that was previously went unnoticed, but
adding a DCHECK unsurfaced this existing issue.

This CL removes that DCHECK since it was recently added.
It is replaced with a TODO so that the underlying issue can be fixed.

Bug: 839892
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I85e486786978f992e2b67a7c67b626fbfe97d898
Reviewed-on: https://chromium-review.googlesource.com/1044506
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#556089}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 939bc11c80dd42f235c126fc515d3db5b7c49b49