Always unfocus omnibox when displaying QRScanner.
On iOS10, when dismissing the keyboard, there is a short time when the
scroll view containing the suggestion is offset.
This cause the omnibox to not be unfocused automatically because of the
test in
https://cs.chromium.org/chromium/src/ios/chrome/browser/ui/content_suggestions/content_suggestions_header_view_controller.mm?l=516
Fix the issue by always unfocus the omnibox when displaying QR scanner.

Bug: 842072
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Ie94884a18335753e919c0cd974c26e4a53abac8b
Reviewed-on: https://chromium-review.googlesource.com/1055394
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#558259}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e53c9ee5b3fce6d95aa91b33ff1af4b7fb1a939d