Show cancel button when the omnibox is focused
This CL shows the cancel button when the omnibox is focused on portrait.
The omnibox can be focused and not the first responder when the user
taps the omnibox, display some suggestions, scroll the suggestions.
The correct check to know if the omnibox is focused is if it is showing
the popup or if it is first responder.

Bug: 836814
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I2d021154f033c81ddde731163c813eb850a973bd
Reviewed-on: https://chromium-review.googlesource.com/1064113
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#559536}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9327148ca68b5094eee729424dcb03c83b6a2dad