[ios] Provides anchoring information for the Bookmarks context menu.
When the UIRefresh UI is enabled, bookmarks is presented using a
UIPresentationController with a custom presentation.  UIAlertController chooses
to draw as a popover in Regular horizontal size classes, so we need to provide a
sourceView and sourceRect.  For now, provide the center of the bottom UIToolbar,
as that roughly corresponds to the "More" button.

BUG=None

Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I6e732460eb371158ac9e1754d752b923965f5b0b
Reviewed-on: https://chromium-review.googlesource.com/1064916
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#560204}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cceb2b29d21849fc705d37e4c77285badee95086