[ios] Add test to update offscreen cells in tab grid
A crash will occur when an scrolled offscreen cell is updated.
This is because UICollectionView's |-cellForItemAtIndexPath:| returns
nil if the cell is offscreen.

This CL adds a unittest that fails without the fix, which is also
in this CL. The fix is to skip cell configuration if the cell is nil.

Bug: 851352
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Iccdb53e3a67b58f225a09ce32ac70f48d42431ef
Reviewed-on: https://chromium-review.googlesource.com/1094210
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#566733}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5c1bb6af234e725a31d311c1237226ba0b78d0d9