[ios] Enable recently closed in tab grid
This CL enables the user to:
1) See recently closed tabs in the tab grid.
2) Tap and open recently close tabs.

This CL adds a method to UrlLoader protocol that restores a tab to
the WebStateList based on SessionID.

In BVC, the restored tab replaces the current WebState.
In tab grid, the restored tab is added to the end of the WebStateList.

Bug: 851219
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I6dc6c22f43d746a9dc73d15b65d6fd5193899a25
Reviewed-on: https://chromium-review.googlesource.com/1094324
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#567175}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d4b94246e31a5504128112fa3f35758b62086b4