[iOS] Don't show selection change when leaving tab grid.
Per the attached bug, this CL changes the tab grid to not show the changed selection state when leaving the tab grid via a tap on a cell.

Cell selection actually happened twice when a cell was tapped; once directly from the collection view, and a second time after the selection had round-tripped through the mediator. This CL cleans that up, and has the collection view decline all selection events, allowing them only to happen as a result of consumer calls.

This CL adds a |showSelectionState| property to the grid view controller; when set to NO, the grid view controller will not update the selection state of the cells when the -selectWithItemID: consumer call is received. The ID of the selected item is maintained, however.

This Cl also doesn't show the selection ring on the proxy cells animated for the transition.

Bug: 849642

Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Ibf8d2c91dc80c42dea55c7e457efe7232d234267
Reviewed-on: https://chromium-review.googlesource.com/1096763
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#567192}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 040c114de8817d1a6f9495e95949f3ee9bf40e66